create enableDataZoom function, move to new LineChart/utils.ts
